requirements.txt
ultralytics
opencv-python
numpy
torch
torchvision
Pillow
matplotlib
pandas
fastapi
uvicorn
python-multipart
easyocr
setuptools
twine
wheel
transformers
python-dotenv
tqdm
joblib
python-dotenv
google-generativeai
pdf2image
openpyxl
streamlit
openai>=1.0.0
json2excel
xlswriter
